工单逻辑更新 by weschen
* 工单逻辑更新

+++ MES-MonitoringService/Model/JobOrder.cs
@@ -43,7 +43,7 @@ namespace MES_MonitoringService.Model
 
 
         [BsonElement("MouldStandardProduceSecond")]
-        public int MouldStandardProduceSecond { get; set; }
+        public decimal MouldStandardProduceSecond { get; set; }
